# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Proxy {
  id: string;
  url: string;
  label: string;
  isActive: boolean;
  successCount: number;
  failCount: number;
  lastUsedAt?: string;
  lastError?: string;
  createdAt: string;
}

export interface BulkProxyInput {
  /** Proxy URLs one per line */
  lines: string;
}

export interface BulkProxyResult {
  added: number;
  skipped: number;
}

export interface AddProxyInput {
  /** Proxy URL e.g. http://user:pass@host:port or socks5://host:port */
  url: string;
  label?: string;
}

export interface ProxyRefreshStatus {
  lastRefreshedAt?: string;
  nextRefreshAt?: string;
  lastAddedCount: number;
  isRefreshing: boolean;
}

export interface RefreshProxiesResult {
  added: number;
  skipped: number;
  total: number;
  lastRefreshedAt?: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface JobCreated {
  jobId: string;
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Job {
  jobId: string;
  status: JobStatus;
  totalRows: number;
  processedRows: number;
  foundCount: number;
  createdAt: string;
  updatedAt: string;
  fileName: string;
}

export type SkipTraceResultStatus = typeof SkipTraceResultStatus[keyof typeof SkipTraceResultStatus];


export const SkipTraceResultStatus = {
  found: 'found',
  not_found: 'not_found',
  error: 'error',
  pending: 'pending',
  processing: 'processing',
} as const;

export interface SkipTraceResult {
  rowIndex: number;
  firstName: string;
  lastName: string;
  address: string;
  city?: string;
  state?: string;
  phones: string[];
  status: SkipTraceResultStatus;
  retryCount: number;
  error?: string;
}

export type JobDetailStatus = typeof JobDetailStatus[keyof typeof JobDetailStatus];


export const JobDetailStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type JobDetailColumnMap = {
  firstNameCol: string;
  lastNameCol: string;
  addressCol: string;
  cityCol?: string;
  stateCol?: string;
};

export interface JobDetail {
  jobId: string;
  status: JobDetailStatus;
  totalRows: number;
  processedRows: number;
  foundCount: number;
  createdAt: string;
  updatedAt: string;
  fileName: string;
  results: SkipTraceResult[];
  columnMap: JobDetailColumnMap;
}

export type UploadSkipTraceListParams = {
/**
 * Column name for first name
 */
firstNameCol?: string;
/**
 * Column name for last name
 */
lastNameCol?: string;
/**
 * Column name for address
 */
addressCol?: string;
/**
 * Column name for city
 */
cityCol?: string;
/**
 * Column name for state
 */
stateCol?: string;
};

export type ListJobs200 = {
  jobs: Job[];
};

export type ListProxies200 = {
  proxies: Proxy[];
};

export type DeleteProxy200 = {
  ok: boolean;
};

export type TestProxy200 = {
  ok: boolean;
  latencyMs?: number;
  error?: string;
};