# lib/api-zod/src/generated/types/refreshProxiesResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshProxiesResult {
  added: number;
  skipped: number;
  total: number;
  lastRefreshedAt?: Date;
}